+++ CHANGELOG.md
@@ -88,7 +88,8 @@
   - Users can now directly upload profile pictures
 - **Reviewer Changes**
   - Simplified voting on agreements as a reviewer
-  - Reviewers can now upload supporting documents to a data request
+  - Reviewers can upload supporting documents to a data request
+  - Reviewers can mark specific data request fields for resubmission
 - **Gem Changes**
   - Updated to ruby 2.4.2
   - Updated to rails 5.1.4